moved messageControls to slice
