--- src/features/messageControls/messageControlsActions.js
@@ -1,2 +1,0 @@
-export const showMessageControls = messageId => ({ type: "messageControls/show", messageId });
-export const hideMessageControls = () => ({ type: "messageControls/hide" });
