--- src/features/messageControls/messageControlsReducer.js
@@ -1,14 +1,0 @@
-const handlers = {
-	"messageControls/show": (state, action) => ({
-		...state,
-		messageId: action.messageId
-	}),
-	"messageControls/hide": state => ({
-		...state,
-		messageId: null
-	})
-};
-
-export default function(state = {}, action) {
-	return handlers[action.type] ? handlers[action.type](state, action) : state;
-}
